requirements.txt
flask==2.3.3
python-magic==0.4.27
cryptography==41.0.3
pyyaml